+++ lib/app/pages/home/widgets/shopping_lists_display.dart
@@ -4,14 +4,11 @@ import 'package:merkar/app/core/resources/app_styles.dart';
 import 'package:merkar/app/core/resources/constants.dart';
 import 'package:merkar/app/core/resources/strings.dart';
 import 'package:merkar/app/pages/shopping/shopping_list/shopping_list_page.dart';
-import 'package:merkar/app/pages/shopping/shopping_list/shopping_list_view_model.dart';
 import 'package:merkar/data/entities/shopping_list.dart';
 
-import '../../../../injection_container.dart';
-
-ShoppingListViewModel viewModel = serviceLocator<ShoppingListViewModel>();
-Widget shoppingListsDisplay(List<ShoppingList> list) {
-  if (list.length == 0) {
+Widget shoppingListsDisplay(
+    List<ShoppingList> list, final ValueChanged<int> onRemoveItem) {
+  if (list.isEmpty) {
     return SliverFillRemaining(
       child: Padding(
         padding: const EdgeInsets.all(Constant.normalspace),
@@ -19,18 +16,15 @@ Widget shoppingListsDisplay(List<ShoppingList> list) {
       ),
     );
   }
-  return  listShoppingList(list);
-}
-
-Widget listShoppingList(List<ShoppingList> list) {
   return SliverList(
     delegate: SliverChildBuilderDelegate(
-          (BuildContext context, int index) {
+      (BuildContext context, int index) {
         return Dismissible(
           child: Padding(
             padding: const EdgeInsets.all(10.0),
             child: Container(
-              decoration: AppStyles.listDecoration(index.toDouble()/list.length),
+              decoration:
+                  AppStyles.listDecoration(index.toDouble() / list.length),
               height: 75.0,
               child: ListTile(
                 title: Text(list[index].name!.capitalize()),
@@ -45,34 +39,29 @@ Widget listShoppingList(List<ShoppingList> list) {
               ),
             ),
           ),
-          background: Container(color: Colors.red,child: Icon(Icons.cancel),),
+          background: Container(
+            color: Colors.red,
+            child: Icon(Icons.cancel),
+          ),
           key: Key(list[index].id!),
-          onDismissed: (direction){
-             Scaffold
-                .of(context)
-                .showSnackBar(SnackBar(content: Text("$list[index].name Eliminada")));
-            viewModel.removeList(list[index]);
-            list.removeAt(index);
-            viewModel.notifyListeners();
-          },
+          onDismissed: (direction) => {onRemoveItem(index)},
           confirmDismiss: (DismissDirection direction) async {
             return await showDialog(
               context: context,
               builder: (BuildContext context) {
                 return AlertDialog(
                   shape: AppStyles.borderRadiusDialog,
-                 // contentPadding: EdgeInsets.only(top: 10.0),
+                  // contentPadding: EdgeInsets.only(top: 10.0),
                   title: Center(child: const Text(Strings.confirm)),
                   content: const Text("Estás seguro de eliminar el Elemento?"),
                   actions: <Widget>[
-                    FlatButton(
-                        onPressed: () => Navigator.of(context).pop(true),
-                        child: const Text(Strings.delete)
-                    ),
-                    FlatButton(
+                    TextButton(
                       onPressed: () => Navigator.of(context).pop(false),
                       child: const Text(Strings.calcel),
                     ),
+                    TextButton(
+                        onPressed: () => Navigator.of(context).pop(true),
+                        child: const Text(Strings.delete)),
                   ],
                 );
               },
