+++ lib/injection_container.dart
@@ -13,7 +13,6 @@ import 'app/pages/login/sign_in/login_view_model.dart';
 import 'app/pages/products/new_product/create_new_product_view_model.dart';
 import 'app/pages/purchases/purchase_history/purchase_history_view_model.dart';
 import 'app/pages/purchases/purchase_history_show_info/purchase_history_show_info_view_model.dart';
-import 'app/pages/shopping/new_shopping_list/new_shopping_list_view_model.dart';
 import 'app/pages/shopping/select_my_products/select_my_products_view_model.dart';
 import 'app/pages/shopping/shopping_list/shopping_list_view_model.dart';
 import 'data/local/local_data_source.dart';
@@ -82,9 +81,6 @@ void createViewModels() {
   serviceLocator.registerFactory(() => SelectMyFavoritesViewModel(
         productsRepository: serviceLocator(),
       ));
-
-  serviceLocator.registerFactory(
-      () => NewShoppingListViewModel(repository: serviceLocator()));
 
   serviceLocator.registerFactory(() => SelectMyProductsViewModel(
       shoppingListRepository: serviceLocator(),
